WholePerformanceクラスを別ファイルに移行する
